+++ src/main/java/org/sonar/plugins/delphi/symbol/resolve/NameResolver.java
@@ -738,13 +738,21 @@ public class NameResolver {
   }
 
   private void handleArgumentList(ArgumentListNode node) {
+    if (handleExplicitArrayConstructorInvocation(node)) {
+      return;
+    }
+    disambiguateArguments(node.getArguments(), true);
+  }
+
+  private boolean handleExplicitArrayConstructorInvocation(ArgumentListNode node) {
     Node previous = node.jjtGetParent().jjtGetChild(node.jjtGetChildIndex() - 1);
     if (previous instanceof NameReferenceNode
         && isExplicitArrayConstructorInvocation(((NameReferenceNode) previous))) {
       updateType(((ClassReferenceType) currentType).classType());
-      return;
+      node.getArguments().forEach(NameResolutionUtils::resolve);
+      return true;
     }
-    disambiguateArguments(node.getArguments(), true);
+    return false;
   }
 
   private boolean handleHardTypeCast(List<ExpressionNode> argumentExpressions) {